game as is
Built using the GODOT game engine:
Tiles load in procedurally as Texture buttons. The default texture is a hexagon. 
Future:
When other buttons are pressed this trigers a signal which changes the "potential_State" from "none" to the selected state. then when clicking a texture button the loaded texture changes and the "potential_State" changes to none. 
Reason for delay:
i had to learn the engine, i orginaly built everything using sprites instead of buttons as all tutorials i was relying on used this. but i relaised later that i needed to use texture buttons which load differnetly so i had to restart. I was also ill for 2 weeks